Issue #8: Reorganise folders
